app/requirements.txt
pandas==1.3.4
psycopg2-binary==2.9.3
